# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BrewAsset POS & WMS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events/{id}/staff:
    post:
      operationId: toggleEventStaff
      tags: [events]
      summary: Assign or remove staff from event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ToggleStaffBody"
      responses:
        "200":
          description: Staff updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /transfers:
    get:
      operationId: getTransfers
      tags: [transfers]
      summary: Get transfers
      parameters:
        - name: repUsername
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Transfer list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Transfer"
    post:
      operationId: createTransfer
      tags: [transfers]
      summary: Create a transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransferBody"
      responses:
        "200":
          description: Transfer created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transfer"
  /staff:
    get:
      operationId: getPromoStaff
      tags: [staff]
      summary: Get promo staff
      responses:
        "200":
          description: Staff list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PromoStaff"
    post:
      operationId: createPromoStaff
      tags: [staff]
      summary: Add promo staff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromoStaffBody"
      responses:
        "200":
          description: Staff created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromoStaff"
  /staff/{id}:
    put:
      operationId: updatePromoStaff
      tags: [staff]
      summary: Update promo staff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePromoStaffBody"
      responses:
        "200":
          description: Staff updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PromoStaff"
    delete:
      operationId: deletePromoStaff
      tags: [staff]
      summary: Remove promo staff
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Staff removed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    ToggleStaffBody:
      type: object
      properties:
        staffName:
          type: string
      required: [staffName]
    Transfer:
      type: object
      properties:
        id:
          type: integer
        repUsername:
          type: string
        itemType:
          type: string
        brand:
          type: string
        quantity:
          type: integer
        fromAccount:
          type: string
        toAccount:
          type: string
        isReturn:
          type: boolean
        createdAt:
          type: string
      required: [id, repUsername, itemType, brand, quantity, fromAccount, toAccount, isReturn, createdAt]
    CreateTransferBody:
      type: object
      properties:
        itemType:
          type: string
        brand:
          type: string
        quantity:
          type: integer
        fromAccount:
          type: string
        toAccount:
          type: string
      required: [itemType, brand, quantity, fromAccount, toAccount]
    PromoStaff:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [id, name]
    CreatePromoStaffBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
      required: [name]
...