# lib/api-zod/src/generated/types/transfer.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * BrewAsset POS & WMS API
 * OpenAPI spec version: 0.1.0
 */

export interface Transfer {
  id: number;
  repUsername: string;
  itemType: string;
  brand: string;
  quantity: number;
  fromAccount: string;
  toAccount: string;
  isReturn: boolean;
  createdAt: string;
}